+++ frontend/.editorconfig
@@ -1,0 +1,2 @@
+[*.tsx]
+indent_size = 2
