+++ frontend/src/index.sass
@@ -38,3 +38,10 @@ label
   min-height: 40px
 
   @apply flex items-center justify-center justify-self-center px-2 rounded-full  p-3
+
+
+.stretch-item
+  @apply flex items-center justify-between
+
+  :first-child
+    @apply font-bold
