+++ frontend/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "baseUrl": ".",
     "strict": true,
     "target": "ESNext",
     "module": "ESNext",
@@ -11,6 +12,9 @@
     "types": [
       "vite/client"
     ],
+    "paths": {
+      "~/*": ["src/*"]
+    },
     "noEmit": true,
     "isolatedModules": true
   }
